ai_resume_builder/requirements.txt
Flask
azure-ai-textanalytics
azure-storage-blob
python-docx
python-dotenv
authlib
requests
azure-ai-formrecognizer 
azure-ai-textanalytics